LPS-140994 Move and rename sxp_elements js
